feat(zsh): cleanup

+++ zshrc/.zshrc_custom/alias.zsh
@@ -1,8 +1,6 @@
 #
 # User configuration
 # New Alias
-alias p="cd ~/Projects"
-alias b="cd ~/blog"
 alias myip='ifconfig | grep "inet " | grep -v 127.0.0.1'
 alias getNodeSize='find . -name "node_modules" -type d -prune -print | xargs du -chs'
 
@@ -19,10 +17,13 @@ alias ...="cd ../.."
 alias ....="cd ../../.."
 
 # Usefull stuff for presentation and seeing dotfiles
-alias hidedesktop="defaults write com.apple.finder CreateDesktop -bool false && killall Finder"
-alias showdesktop="defaults write com.apple.finder CreateDesktop -bool true && killall Finder"
-alias showall='defaults write com.apple.finder AppleShowAllFiles YES && killall Finder'
-alias hideall='defaults write com.apple.finder AppleShowAllFiles NO && killall Finder'
+#
+if [[ "$(uname)" == "Darwin" ]]; then
+    alias hidedesktop="defaults write com.apple.finder CreateDesktop -bool false && killall Finder"
+    alias showdesktop="defaults write com.apple.finder CreateDesktop -bool true && killall Finder"
+    alias showall='defaults write com.apple.finder AppleShowAllFiles YES && killall Finder'
+    alias hideall='defaults write com.apple.finder AppleShowAllFiles NO && killall Finder'
+fi
 
 # Get rid of those pesky .DS_Store files recursively
 alias dsclean='find . -type f -name .DS_Store -print0 | xargs -0 rm'
@@ -30,26 +31,26 @@ alias dsclean='find . -type f -name .DS_Store -print0 | xargs -0 rm'
 #
 # Git, run inside the git project will change directory to the git root
 #
-alias githome="cd $(git rev-parse --show-toplevel)"
-
-#
-# NeoVide
-#
-# Need to be installed and configure
-# alias neovide=~/Github/neovide/target/release/neovide --multigrid
+function gitroot() {
+  local root
+  root=$(git rev-parse --show-toplevel 2>/dev/null)
+  if [[ -n "$root" ]]; then
+    cd "$root"
+  else
+    echo "Not in a git repository"
+  fi
+}
 
 #
 # Help me setup the correct name for git repository between personal and work
 #
-# 
-alias personal_git="git config user.email 'bozhidar.dryanovski@gmail.com' && git config user.name 'Bozhidar Dryanovski' && echo 'All Done'"
+if [[ -v EMAIL ]] && [[ -v FULLNAME ]]; then
+    alias personal_git="git config user.email '$EMAIL' && git config user.name '$FULLNAME' && echo 'All Done'"
+fi
 
 #
-# Configure to use bat for cat files 
+# Configure to use bat for cat files
 #
 if command -v bat >/dev/null 2>&1; then
   alias cat="bat --theme DarkNeon"
 fi
-
-
-alias mysql=/opt/homebrew/opt/mysql@8.0/bin/mysql
